fix: unified switch style
