+++ src/locales/fa.json
@@ -261,7 +261,7 @@
   "Restart": "راه‌اندازی مجدد",
   "Release Version": "نسخه نهایی",
   "Alpha Version": "نسخه آلفا",
-  "Tun mode requires": "Tun mode نیاز دارد",
+  "Please Install and Enable Service Mode First": "لطفاً ابتدا حالت سرویس را نصب و فعال کنید",
   "Grant": "اعطا",
   "Open UWP tool": "باز کردن ابزار UWP",
   "Open UWP tool Info": "از ویندوز 8 به بعد، برنامه‌های UWP (مانند Microsoft Store) از دسترسی مستقیم به خدمات شبکه محلی محدود شده‌اند و این ابزار می‌تواند برای دور زدن این محدودیت استفاده شود",
