+++ src/locales/ru.json
@@ -264,7 +264,7 @@
   "Restart": "Перезапуск",
   "Release Version": "Официальная версия",
   "Alpha Version": "Альфа-версия",
-  "Tun mode requires": "Требуется Режим туннеля",
+  "Please Enable Service Mode": "Пожалуйста, сначала установите и включите режим обслуживания",
   "Grant": "Предоставить",
   "Open UWP tool": "Открыть UWP инструмент",
   "Open UWP tool Info": "С Windows 8 приложения UWP (такие как Microsoft Store) ограничены в прямом доступе к сетевым службам локального хоста, и этот инструмент позволяет обойти это ограничение",
